refactor: consolidate duplicated zero controllers and tanh_sigmoid
Add shared zero_controller(n_controls) factory in controllers/zero.py.
System-specific zero controllers now delegate to it instead of each
having 30+ lines of identical boilerplate.

Consolidate duplicated tanh_sigmoid — miscellaneous.py now re-exports
from real_functions.py (the JIT'd version with configurable k).
