+++ src/cbfkit/systems/quadrotor_6dof/controllers/zero.py
@@ -1,48 +1,8 @@
-import jax.numpy as jnp
-from jax import Array, jit
+"""Zero controller for quadrotor (4 control inputs)."""
 
-from cbfkit.utils.user_types import (
-    Control,
-    ControllerCallableReturns,
-    ControllerData,
-    Key,
-    NominalControllerCallable,
-    Optional,
-    State,
-)
+from cbfkit.controllers.zero import zero_controller as _zero
+from cbfkit.utils.user_types import NominalControllerCallable
 
 
 def zero_controller() -> NominalControllerCallable:
-    """Create a zero controller for the given quadrotor dynamics.
-
-    Args:
-        None
-
-    Returns
-    -------
-        controller (Callable): handle to function computing zero control
-    """
-
-    @jit
-    def controller(
-        _t: float, _state: State, _key: Key, _xd: Optional[State] = None
-    ) -> ControllerCallableReturns:
-        """Computes zero control input (4x1).
-
-        Args:
-            _t (float): time in sec
-            _state (Array): state vector (or estimate if using observer/estimator)
-            _key (Key): unused
-            _xd (Optional[State]): unused
-
-        Returns
-        -------
-            zeros (Array): 4x1 zero vector
-            data: ControllerData: empty ControllerData
-        """
-        # logging data
-        data = ControllerData()
-
-        return jnp.zeros((4,)), data
-
-    return controller
+    return _zero(n_controls=4)
